refactor: remove unnecessary RecordProcessor class
