+++ airbyte/_future_cdk/sql_processor.py
@@ -6,6 +6,7 @@ from __future__ import annotations
 import abc
 import contextlib
 import enum
+from collections import defaultdict
 from contextlib import contextmanager
 from functools import cached_property
 from pathlib import Path
@@ -29,8 +30,17 @@ from sqlalchemy import (
 )
 from sqlalchemy.sql.elements import TextClause
 
+from airbyte_protocol.models import (
+    AirbyteMessage,
+    AirbyteRecordMessage,
+    AirbyteStateMessage,
+    AirbyteStateType,
+    AirbyteStreamState,
+    AirbyteTraceMessage,
+    Type,
+)
+
 from airbyte import exceptions as exc
-from airbyte._future_cdk.record_processor import RecordProcessorBase
 from airbyte._future_cdk.state_writers import StdOutStateWriter
 from airbyte._util.name_normalizers import LowerCaseNormalizer
 from airbyte.constants import (
@@ -39,12 +49,13 @@ from airbyte.constants import (
     AB_RAW_ID_COLUMN,
     DEBUG_MODE,
 )
+from airbyte.records import StreamRecordHandler
 from airbyte.strategies import WriteStrategy
 from airbyte.types import SQLTypeConverter
 
 
 if TYPE_CHECKING:
-    from collections.abc import Generator
+    from collections.abc import Generator, Iterable
 
     from sqlalchemy.engine import Connection, Engine
     from sqlalchemy.engine.cursor import CursorResult
@@ -52,17 +63,11 @@ if TYPE_CHECKING:
     from sqlalchemy.sql.base import Executable
     from sqlalchemy.sql.type_api import TypeEngine
 
-    from airbyte_protocol.models import (
-        AirbyteRecordMessage,
-        AirbyteStateMessage,
-    )
-
     from airbyte._batch_handles import BatchHandle
     from airbyte._future_cdk.catalog_providers import CatalogProvider
     from airbyte._future_cdk.state_writers import StateWriterBase
     from airbyte._processors.file.base import FileWriterBase
     from airbyte.progress import ProgressTracker
-    from airbyte.records import StreamRecordHandler
     from airbyte.secrets.base import SecretString
 
 
@@ -116,7 +121,7 @@ class SqlConfig(BaseModel, abc.ABC):
         )
 
 
-class SqlProcessorBase(RecordProcessorBase):
+class SqlProcessorBase(abc.ABC):
     """A base class to be used for SQL Caches."""
 
     type_converter_class: type[SQLTypeConverter] = SQLTypeConverter
@@ -130,8 +135,6 @@ class SqlProcessorBase(RecordProcessorBase):
 
     supports_merge_insert = False
     """True if the database supports the MERGE INTO syntax."""
-
-    # Constructor:
 
     def __init__(
         self,
@@ -152,10 +155,16 @@ class SqlProcessorBase(RecordProcessorBase):
 
         self._sql_config: SqlConfig = sql_config
 
-        super().__init__(
-            state_writer=state_writer,
-            catalog_provider=catalog_provider,
-        )
+        self._catalog_provider: CatalogProvider | None = catalog_provider
+        self._state_writer: StateWriterBase | None = state_writer or StdOutStateWriter()
+
+        self._pending_state_messages: dict[str, list[AirbyteStateMessage]] = defaultdict(list, {})
+        self._finalized_state_messages: dict[
+            str,
+            list[AirbyteStateMessage],
+        ] = defaultdict(list, {})
+
+        self._setup()
         self.file_writer = file_writer or self.file_writer_class(
             cache_dir=cast(Path, temp_dir),
             cleanup=temp_file_cleanup,
@@ -165,6 +174,155 @@ class SqlProcessorBase(RecordProcessorBase):
 
         self._known_schemas_list: list[str] = []
         self._ensure_schema_exists()
+
+    # Inherited methods
+
+    # @property
+    # def expected_streams(self) -> set[str]:
+    #     """Return the expected stream names."""
+    #     return set(self.catalog_provider.stream_names)
+
+    @property
+    def catalog_provider(
+        self,
+    ) -> CatalogProvider:
+        """Return the catalog manager.
+
+        Subclasses should set this property to a valid catalog manager instance if one
+        is not explicitly passed to the constructor.
+
+        Raises:
+            PyAirbyteInternalError: If the catalog manager is not set.
+        """
+        if not self._catalog_provider:
+            raise exc.PyAirbyteInternalError(
+                message="Catalog manager should exist but does not.",
+            )
+
+        return self._catalog_provider
+
+    @property
+    def state_writer(
+        self,
+    ) -> StateWriterBase:
+        """Return the state writer instance.
+
+        Subclasses should set this property to a valid state manager instance if one
+        is not explicitly passed to the constructor.
+
+        Raises:
+            PyAirbyteInternalError: If the state manager is not set.
+        """
+        if not self._state_writer:
+            raise exc.PyAirbyteInternalError(
+                message="State manager should exist but does not.",
+            )
+
+        return self._state_writer
+
+    @final
+    def process_airbyte_messages(
+        self,
+        messages: Iterable[AirbyteMessage],
+        *,
+        catalog_provider: CatalogProvider | None = None,
+        write_strategy: WriteStrategy,
+        progress_tracker: ProgressTracker,
+    ) -> None:
+        """Process a stream of Airbyte messages."""
+        if not isinstance(write_strategy, WriteStrategy):
+            raise exc.AirbyteInternalError(
+                message="Invalid `write_strategy` argument. Expected instance of WriteStrategy.",
+                context={"write_strategy": write_strategy},
+            )
+
+        stream_record_handlers: dict[str, StreamRecordHandler] = {}
+
+        # Process messages, writing to batches as we go
+        for message in messages:
+            if message.type is Type.RECORD:
+                record_msg = cast(AirbyteRecordMessage, message.record)
+                stream_name = record_msg.stream
+
+                if stream_name not in stream_record_handlers:
+                    stream_record_handlers[stream_name] = StreamRecordHandler(
+                        json_schema=self.catalog_provider.get_stream_json_schema(
+                            stream_name=stream_name,
+                        ),
+                        normalize_keys=True,
+                        prune_extra_fields=True,
+                    )
+
+                self.process_record_message(
+                    record_msg,
+                    stream_record_handler=stream_record_handlers[stream_name],
+                    progress_tracker=progress_tracker,
+                )
+
+            elif message.type is Type.STATE:
+                state_msg = cast(AirbyteStateMessage, message.state)
+                if state_msg.type in {AirbyteStateType.GLOBAL, AirbyteStateType.LEGACY}:
+                    self._pending_state_messages[f"_{state_msg.type}"].append(state_msg)
+                else:
+                    stream_state = cast(AirbyteStreamState, state_msg.stream)
+                    stream_name = stream_state.stream_descriptor.name
+                    self._pending_state_messages[stream_name].append(state_msg)
+
+            elif message.type is Type.TRACE:
+                trace_msg: AirbyteTraceMessage = cast(AirbyteTraceMessage, message.trace)
+                if trace_msg.stream_status and trace_msg.stream_status.status == "SUCCEEDED":
+                    # This stream has completed successfully, so go ahead and write the data.
+                    # This will also finalize any pending state messages.
+                    self.write_stream_data(
+                        stream_name=trace_msg.stream_status.stream_descriptor.name,
+                        write_strategy=write_strategy,
+                        progress_tracker=progress_tracker,
+                    )
+
+            else:
+                # Ignore unexpected or unhandled message types:
+                # Type.LOG, Type.CONTROL, etc.
+                pass
+
+        # We've finished processing input data.
+        # Finalize all received records and state messages:
+        self.write_all_stream_data(
+            write_strategy=write_strategy,
+            progress_tracker=progress_tracker,
+        )
+
+        self.cleanup_all()
+
+    def write_all_stream_data(
+        self,
+        write_strategy: WriteStrategy,
+        progress_tracker: ProgressTracker,
+    ) -> None:
+        """Finalize any pending writes."""
+        for stream_name in self.catalog_provider.stream_names:
+            self.write_stream_data(
+                stream_name,
+                write_strategy=write_strategy,
+                progress_tracker=progress_tracker,
+            )
+
+    def _finalize_state_messages(
+        self,
+        state_messages: list[AirbyteStateMessage],
+    ) -> None:
+        """Handle state messages by passing them to the catalog manager."""
+        if state_messages:
+            self.state_writer.write_state(
+                state_message=state_messages[-1],
+            )
+
+    def _setup(self) -> None:  # noqa: B027  # Intentionally empty, not abstract
+        """Create the database.
+
+        By default this is a no-op but subclasses can override this method to prepare
+        any necessary resources.
+        """
+        pass
 
     # Public interface:
